refactor(storage): split remote driver into submodules
Split the monolithic `src/storage/drivers/remote.rs` (885 lines) into focused submodules:
- `list.rs`: list_paths implementation
- `multipart.rs`: multipart upload operations and HashingReader
- `presigned.rs`: presigned URL generation
- `storage_driver.rs`: core StorageDriver trait implementation
- `stream_upload.rs`: streaming upload operations
- `tests.rs`: all unit and integration tests
- `mod.rs`: shared types and driver construction

Each submodule now handles a single storage capability. The public API and test coverage remain unchanged.
